revision: refactor catDecorator for decoupling
